# test/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/system/getDatabaseInfo/option": {
            "get": {
                "operationId": "System_getDatabaseInfoOption",
                "summary": "获取数据库信息 option",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Option"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "system"
                ]
            }
        },
        "/api/system/getDatabaseInfo": {
            "get": {
                "operationId": "System_getDatabaseInfo",
                "summary": "获取数据库信息",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DatabaseInfoDto"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "system"
                ]
            }
        },
        "/api/system/getOsInfo/option": {
            "get": {
                "operationId": "System_getOsInfoOption",
                "summary": "获取系统信息 option",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Option"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "system"
                ]
            }
        },
        "/api/system/getOsInfo": {
            "get": {
                "operationId": "System_getOsInfo",
                "summary": "获取系统信息",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OsInfoDto"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "创建成功"
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "system"
                ]
            }
        },
        "/api/system/sqliteVacuum": {
            "post": {
                "operationId": "System_sqliteVacuum",
                "summary": "释放 sqlite 数据库未使用空间",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "响应成功"
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "system"
                ]
            }
        },
        "/api/system/disableEmptyFeeds": {
            "post": {
                "operationId": "System_disableEmptyFeeds",
                "summary": "禁用不包含任何 Hook 和 自定义查询的订阅",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "响应成功"
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "数据格式有误"
                    },
                    "401": {
                        "description": "没有权限访问"
                    },
                    "403": {
                        "description": "该用户没有权限访问"
                    },
                    "404": {
                        "description": "找不到要访问的接口"
                    },
                    "408": {
                        "description": "请求超时"
                    },
                    "409": {
                        "description": "请求时发生了冲突"
                    },
                    "429": {
                        "description": "请求次数超限"
                    },
                    "500": {
                        "description": "服务器内部错误"
                    },
                    "502": {
                        "description": "代理服务器错误"
                    }
                },
                "tags": [
                    "system"
                ]
            }
        },
...
    },
    "info": {
        "title": "RSS Impact server docs",
        "description": "RSS Impact server docs",
        "version": "0.1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            }
        },
        "schemas": {
            "ResponseDto": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    },
                    "stack": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
            "Rule": {
                "type": "object",
                "properties": {
                    "required": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "trigger": {
                        "type": "string"
                    }
                },
                "required": [
                    "required",
                    "message",
                    "trigger"
                ]
            },
            "DicData": {
                "type": "object",
                "properties": {
                    "label": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                },
                "required": [
                    "label",
                    "value"
                ]
            },
            "Column": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "rules": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rule"
                        }
                    },
                    "dicData": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DicData"
                        }
                    },
                    "label": {
                        "type": "string"
                    },
                    "prop": {
                        "type": "string"
                    },
                    "overHidden": {
                        "type": "boolean"
                    },
                    "value": {
                        "type": "object"
                    },
                    "addDisplay": {
                        "type": "boolean"
                    },
                    "editDisabled": {
                        "type": "boolean"
                    },
                    "readonly": {
                        "type": "boolean"
                    },
                    "max": {
                        "type": "number"
                    },
                    "min": {
                        "type": "number"
                    },
                    "format": {
                        "type": "string"
                    },
                    "maxlength": {
                        "type": "number"
                    },
                    "minlength": {
                        "type": "number"
                    },
                    "alone": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "type",
                    "label",
                    "prop"
                ]
            },
            "Option": {
                "type": "object",
                "properties": {
                    "column": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Column"
                        }
                    },
                    "index": {
                        "type": "boolean"
                    },
                    "align": {
                        "type": "string"
                    },
                    "border": {
                        "type": "boolean"
                    },
                    "stripe": {
                        "type": "boolean"
                    },
                    "columnBtn": {
                        "type": "boolean"
                    },
                    "refreshBtn": {
                        "type": "boolean"
                    },
                    "addBtn": {
                        "type": "boolean"
                    },
                    "editBtn": {
                        "type": "boolean"
                    },
                    "delBtn": {
                        "type": "boolean"
                    },
                    "viewBtn": {
                        "type": "boolean"
                    },
                    "excelBtn": {
                        "type": "boolean"
                    },
                    "title": {
                        "type": "string"
                    },
                    "addRowBtn": {
                        "type": "boolean"
                    },
                    "calcHeight": {
                        "type": "number"
                    },
                    "cancelBtnTitle": {
                        "type": "string"
                    },
                    "dataType": {
                        "type": "string"
                    },
                    "cellBtn": {
                        "type": "boolean"
                    },
                    "dateBtn": {
                        "type": "boolean"
                    },
                    "cancelBtn": {
                        "type": "boolean"
                    },
                    "dateDefault": {
                        "type": "boolean"
                    },
                    "dicData": {
                        "type": "object"
                    },
                    "dicMethod": {
                        "type": "string"
                    },
                    "dicQuery": {
                        "type": "object"
                    },
                    "dicUrl": {
                        "type": "string"
                    },
                    "defaultSort": {
                        "type": "object"
                    },
                    "dialogFullscreen": {
                        "type": "boolean"
                    },
                    "dialogEscape": {
                        "type": "boolean"
                    },
                    "dialogClickModal": {
                        "type": "boolean"
                    },
                    "dialogCloseBtn": {
                        "type": "boolean"
                    },
                    "dialogModal": {
                        "type": "boolean"
                    },
                    "dialogTop": {
                        "type": "object"
                    },
                    "dialogType": {
                        "type": "string"
                    },
                    "dialogWidth": {
                        "type": "object"
                    },
                    "dialogHeight": {
                        "type": "object"
                    },
                    "defaultExpandAll": {
                        "type": "boolean"
                    },
                    "expandRowKeys": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "emptyText": {
                        "type": "string"
                    },
                    "expand": {
                        "type": "boolean"
                    },
                    "expandWidth": {
                        "type": "number"
                    },
                    "expandFixed": {
                        "type": "boolean"
                    },
                    "filterBtn": {
                        "type": "boolean"
                    },
                    "formWidth": {
                        "type": "object"
                    },
                    "height": {
                        "type": "number"
                    },
                    "header": {
                        "type": "boolean"
                    },
                    "indexLabel": {
                        "type": "string"
                    },
                    "indexWidth": {
                        "type": "number"
                    },
                    "indexFixed": {
                        "type": "boolean"
                    },
                    "rowKey": {
                        "type": "string"
                    },
                    "indeterminate": {
                        "type": "boolean"
                    },
                    "labelWidth": {
                        "type": "number"
                    },
                    "maxHeight": {
                        "type": "number"
                    },
                    "menu": {
                        "type": "boolean"
                    },
                    "menuWidth": {
                        "type": "number"
                    },
                    "menuXsWidth": {
                        "type": "number"
                    },
                    "menuAlign": {
                        "type": "string"
                    },
                    "menuType": {
                        "type": "string"
                    },
                    "menuBtnTitle": {
                        "type": "string"
                    },
                    "pageSize": {
                        "type": "string"
                    },
                    "pageSizes": {
                        "type": "array",
                        "items": {
                            "type": "number"
                        }
                    },
                    "printBtn": {
                        "type": "boolean"
                    },
                    "saveBtn": {
                        "type": "boolean"
                    },
                    "updateBtn": {
                        "type": "boolean"
                    },
                    "cancalBtn": {
                        "type": "boolean"
                    },
                    "saveBtnTitle": {
                        "type": "string"
                    },
                    "selection": {
                        "type": "boolean"
                    },
                    "selectionWidth": {
                        "type": "number"
                    },
                    "selectionFixed": {
                        "type": "boolean"
                    },
                    "searchBtn": {
                        "type": "boolean"
                    },
                    "selectable": {
                        "type": "boolean"
                    },
                    "selectClearBtn": {
                        "type": "boolean"
                    },
                    "showHeader": {
                        "type": "boolean"
                    },
                    "showSummary": {
                        "type": "boolean"
                    },
                    "sumColumnList": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "tip": {
                        "type": "string"
                    },
                    "tipPlacement": {
                        "type": "string"
                    },
                    "checkStrictly": {
                        "type": "boolean"
                    },
                    "updateBtnTitle": {
                        "type": "string"
                    },
                    "width": {
                        "type": "number"
                    }
                },
                "required": [
                    "column"
                ]
            },
...
            "DatabaseInfoDto": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "title": "类型",
                        "description": "参考 typeorm 支持的数据库类型。",
                        "example": "sqlite"
                    },
                    "size": {
                        "type": "number",
                        "title": "体积(B)",
                        "description": "单位为 B",
                        "example": 114514
                    },
                    "sizeFormat": {
                        "type": "string",
                        "title": "体积(B)",
                        "description": "单位为 B",
                        "example": "114.51 MiB"
                    },
                    "entitiesLength": {
                        "type": "number",
                        "title": "实体类数量",
                        "description": "typeorm 的实体类数量",
                        "example": 10
                    },
                    "tableCount": {
                        "type": "number",
                        "title": "表格数量",
                        "description": "SQL 表数量",
                        "example": 10
                    },
                    "indexCount": {
                        "type": "number",
                        "title": "索引数量",
                        "description": "SQL 索引数量",
                        "example": 10
                    }
                },
                "required": [
                    "type",
                    "size",
                    "sizeFormat",
                    "entitiesLength",
                    "tableCount",
                    "indexCount"
                ]
            },
            "OsInfoDto": {
                "type": "object",
                "properties": {
                    "nodeVersion": {
                        "type": "string",
                        "title": "Node.js 版本",
                        "description": "",
                        "example": ""
                    },
                    "hostname": {
                        "type": "string",
                        "title": "主机名",
                        "description": "",
                        "example": ""
                    },
                    "type": {
                        "type": "string",
                        "title": "操作系统类型",
                        "description": "",
                        "example": ""
                    },
                    "platform": {
                        "type": "string",
                        "title": "操作系统平台",
                        "description": "",
                        "example": ""
                    },
                    "arch": {
                        "type": "string",
                        "title": "系统架构",
                        "description": "",
                        "example": ""
                    },
                    "release": {
                        "type": "string",
                        "title": "系统版本",
                        "description": "",
                        "example": ""
                    },
                    "totalmem": {
                        "type": "string",
                        "title": "总内存",
                        "description": "",
                        "example": ""
                    },
                    "freemem": {
                        "type": "string",
                        "title": "可用内存",
                        "description": "",
                        "example": ""
                    },
                    "cpuNum": {
                        "type": "number",
                        "title": "CPU 内核数",
                        "description": "",
                        "example": ""
                    },
                    "osUptime": {
                        "type": "string",
                        "title": "系统运行时间",
                        "description": "",
                        "example": ""
                    },
                    "uptime": {
                        "type": "string",
                        "title": "进程运行时间",
                        "description": "",
                        "example": ""
                    }
                },
                "required": [
                    "nodeVersion",
                    "hostname",
                    "type",
                    "platform",
                    "arch",
                    "release",
                    "totalmem",
                    "freemem",
                    "cpuNum",
                    "osUptime",
                    "uptime"
                ]
            },
...
        }
    }
}